CAN between nodes, Node 2 Servo and LED
CAN between nodes tested by sending a joystick position from Node 1 to Node 2.
Node 2 servo uses the Servo class from Arduino
LED score count should maybe be removed: Rather poll the LED (and other inputs received over CAN, eg joystick), and send a CAN message to Node 1 when the game is over / score has changed.
Node 2 uart code moved to separate file.
Node 2 CAN loopback test needs to be moved/removed.

+++ ATmega162/drivers/analog/joystick.c
@@ -57,7 +57,7 @@ JOY_dir_t JOY_get_direction(void){
                               NEUTRAL;
 }
 
-int JOY_get_button(void){
+int8_t JOY_get_button(void){
 	return !(PINB & (1<<PINB2));
 }
